Extend

+++ Section_7/Animal.java
@@ -3,13 +3,13 @@ package Section_7;
 public class Animal {
     private String type;
     private String size;
-    private int weight;
+    private double weight;
 
     public Animal(){
 
     }
 
-    public Animal(String type, String size, int weight) {
+    public Animal(String type, String size, double weight) {
         this.type = type;
         this.size = size;
         this.weight = weight;
